QCLI Test automation - merged ai_prompts feature test into one
